+++ src/Core/Algorithms/Legacy/FiniteElements/BuildRHS/BuildFEVolRHS.cc
@@ -566,7 +566,7 @@ void FEMVolRHSBuilder::parallel(int proc_num)
 		{
 
                         //zero output vector
-                        const auto ii = static_cast<int>(i);
+                        const auto ii = static_cast<uint64_t>(i);
                         (*rhsmatrix_)(ii,0)=0.0;
 
 			if (i < global_dimension_nodes)
@@ -599,7 +599,7 @@ void FEMVolRHSBuilder::parallel(int proc_num)
 					{
 						if (na[k] == i)
 						{
-              const auto ii = static_cast<int>(i);
+              const auto ii = static_cast<uint64_t>(i);
 					          build_local_matrix_regular(ca[j], k , l_val, ni_points, ni_weights, ni_derivatives,precompute);
                                                   (*rhsmatrix_)(ii,0)+=l_val;
 						}
@@ -624,7 +624,7 @@ void FEMVolRHSBuilder::parallel(int proc_num)
 					{
 						if (na[k] == i)
 						{
-              const auto ii = static_cast<int>(i);
+              const auto ii = static_cast<uint64_t>(i);
 					            build_local_matrix(ca[j], k , l_val, ni_points, ni_weights, ni_derivatives);
                                                     (*rhsmatrix_)(ii,0)+=l_val;  //rhsmatrix_->add(i, 0, l_val);
 						}
@@ -636,7 +636,7 @@ void FEMVolRHSBuilder::parallel(int proc_num)
 						{
 							if (global_dimension + static_cast<int>(ea[k]) == i)
 							{
-                const auto ii = static_cast<int>(i);
+                const auto ii = static_cast<uint64_t>(i);
 							     build_local_matrix(ca[j], k+na.size() , l_val, ni_points, ni_weights, ni_derivatives);
                                                              (*rhsmatrix_)(ii,0)+=l_val;
 							}
